Split up dsbs backend from frontend and so the backend can be used separatly.
